+++ pkg/graph/graph_test.go
@@ -172,3 +172,70 @@ func TestContext(t *testing.T) {
 	}
 
 }
+
+func TestNodeIdAsTag(t *testing.T) {
+	b := NewBuilder()
+
+	type CounterCfg struct {
+		From int
+		To   int
+	}
+	RegisterStart(b, func(cfg CounterCfg) node.StartFuncCtx[int] {
+		return func(_ context.Context, out chan<- int) {
+			for i := cfg.From; i <= cfg.To; i++ {
+				out <- i
+			}
+		}
+	})
+
+	type DoublerCfg struct{}
+	RegisterMiddle(b, func(_ DoublerCfg) node.MiddleFunc[int, int] {
+		return func(in <-chan int, out chan<- int) {
+			for n := range in {
+				out <- n * 2
+			}
+		}
+	})
+
+	type MapperCfg struct {
+		Dst map[int]struct{}
+	}
+	RegisterTerminal(b, func(cfg MapperCfg) node.TerminalFunc[int] {
+		return func(in <-chan int) {
+			for n := range in {
+				cfg.Dst[n] = struct{}{}
+			}
+		}
+	})
+
+	type config struct {
+		Starts CounterCfg `nodeId:"s"`
+		Middle DoublerCfg `nodeId:"m"`
+		Term   MapperCfg  `nodeId:"t"`
+		Connector
+	}
+	map1 := map[int]struct{}{}
+	g, err := b.Build(config{
+		Starts: CounterCfg{From: 1, To: 5},
+		Middle: DoublerCfg{},
+		Term:   MapperCfg{Dst: map1},
+		Connector: Connector{
+			"s": {"m"},
+			"m": {"t"},
+		},
+	})
+	require.NoError(t, err)
+
+	done := make(chan struct{})
+	go func() {
+		g.Run(context.Background())
+		close(done)
+	}()
+	select {
+	case <-done:
+	case <-time.After(5 * time.Second):
+		require.Fail(t, "timeout while waiting for graph to complete")
+	}
+
+	assert.Equal(t, map[int]struct{}{2: {}, 4: {}, 6: {}, 8: {}, 10: {}}, map1)
+}