+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # CHANGELOG
 
-# v0.6.0
+# v0.NEXT
+* High-level API nodes do not need to implement `Instancer` if you define a `nodeId` tag in the
+  config struct that defines them.
+
 ## Breaking changes
 * High-level graph API add contexts to the `graph.Run(context.Context)` library, and
 * Start providers in High-level API must return a `StartFuncCtx[OUT]` function instead of a
   `StartFunc[OUT]`
 * In version 1.0 maybe StartFunc is replaced by StartFuncCtx and we force the usage of contexts always
+
+## TO DO
+* Allow also an `instanceId` field inside the struct so any string or stringer can be marked
+as InstanceId without needing to provide an `ID` implementor.
+* Allow that all nodes sharing a nodeId can treated as a group for sending/receiving in the connector info
 
 # v0.5.0
 * Context propagation. Added: `StartFuncCtx[OUT]` type, `AsStartCtx` function and `StartCtx` method.